+++ gaios/Helpers/Constants.swift
@@ -52,6 +52,7 @@ enum ExternalUrls {
     static let helpReceiveCapacity = "https://help.blockstream.com/hc/en-us/articles/18788499177753"
     static let helpReceiveFees = "https://help.blockstream.com/hc/en-us/articles/18788578831897"
     static let helpRecoveryTransactions = "https://help.blockstream.com/hc/en-us/articles/900004249546-The-upgrade-from-nLockTime-to-CheckSequenceVerify"
+    static let pinServerSupport = "https://help.blockstream.com/hc/en-us/requests/new?tf_900008231623=ios&tf_subject=Non-default%20PIN%20server&&tf_900003758323=blockstream_jade&tf_900006375926=jade&tf_900009625166="
 }
 
 enum RiveModel {